EmoTrack-API/requirements.txt
flask
flask_mail
flask_jwt_extended
python-dotenv
requests
numpy
pandas
seaborn
scikit-learn
joblib
deepface